changing CONFIG_AUTOREMOVE
since January 26th the compilation fails when trying to compile and install the toolchain. The error happens when trying to untar a file that does not exist. Apparently at some point the file is deleted after downloading:

2023-02-22T08:57:50.4510447Z . /workdir/openwrt/include/shell.sh; /workdir/openwrt/staging_dir/host/bin/libdeflate-gzip -dc /workdir/openwrt/dl/musl-1.2.3.tar.gz | tar -C /workdir/openwrt/build_dir/toolchain-aarch64_cortex-a53_gcc-12.2.0_musl/musl-1.2.3/.. -xf - 2023-02-22T08:57:50.4555117Z libdeflate-gzip: Can't open "/workdir/openwrt/dl/musl-1.2.3.tar.gz" for reading: No such file or directory 2023-02-22T08:57:50.4556283Z tar: This does not look like a tar archive

changing from

"CONFIG_AUTOREMOVE=y"

to

"# CONFIG_AUTOREMOVE is not set"

to fix it